pyproject.toml
[project]
name = "agent-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent>=0.1.3",
    "openai-agent>=0.1.0",
    "python-dotenv>=1.1.1",
    "runner>=1.1",
]
